feat(blank-screen): remove AngularJS controller of blank screen view
- Remove AngularJS controller of blank screen view
- Remove usage of controller in the html template
- Move logic to manage the error message to the BlankScreen component
