+++ plugins/main/public/controllers/index.ts
@@ -16,4 +16,3 @@ import './agent';
 import './settings';
 import './security';
 import './tools';
-import './misc';
