+++ plugins/main/public/templates/error-handler/blank-screen.html
@@ -1,5 +1,1 @@
-<div ng-controller="blankScreenController as ctrl">
-    <div ng-if="ctrl.showErrorPage">
-       <react-component name="WzBlankScreen" props="blankScreenProps" ></react-component>
-    </div>
-</div>
+<react-component name="WzBlankScreen"></react-component>
